some refactor, please log in will now try to enable metamask
