+++ src/layouts/HeaderContainer.tsx
@@ -1,4 +1,10 @@
 import { IDAOState } from "@daostack/client";
+import * as uiActions from "actions/uiActions";
+import { enableMetamask, getArc } from "arc";
+import AccountBalances from "components/Account/AccountBalances";
+import AccountImage from "components/Account/AccountImage";
+import AccountProfileName from "components/Account/AccountProfileName";
+import Subscribe, { IObservableState } from "components/Shared/Subscribe";
 import Util from "lib/util";
 import * as queryString from "query-string";
 import * as React from "react";
@@ -6,27 +12,16 @@ import { Breadcrumbs } from "react-breadcrumbs-dynamic";
 import { connect } from "react-redux";
 import { Link, matchPath, NavLink, RouteComponentProps } from "react-router-dom";
 import { IRootState } from "reducers";
-import { IAccountState } from "reducers/arcReducer";
 import { NotificationStatus, showNotification } from "reducers/notifications";
 import { IProfileState } from "reducers/profilesReducer";
-
-import * as uiActions from "actions/uiActions";
-import * as web3Actions from "actions/web3Actions";
-import { getArc } from "arc";
-import AccountBalances from "components/Account/AccountBalances";
-import AccountImage from "components/Account/AccountImage";
-import AccountProfileName from "components/Account/AccountProfileName";
-import Subscribe, { IObservableState } from "components/Shared/Subscribe";
 import * as css from "./App.scss";
 
 interface IStateProps {
   accounts: string[];
   currentAccountProfile: IProfileState;
-  currentAccount: IAccountState;
   dao: IDAOState;
   ethAccountAddress: string | null;
   networkId: number;
-  pageURL: string;
 }
 
 const mapStateToProps = (state: IRootState, ownProps: any) => {
@@ -34,7 +29,6 @@ const mapStateToProps = (state: IRootState, ownProps: any) => {
   return {
     dao,
     ethAccountAddress: state.web3.ethAccountAddress,
-    pageURL: ownProps.location.pathname
   };
 };
 
@@ -70,19 +64,18 @@ class HeaderContainer extends React.Component<IProps, null> {
     showTour();
   }
 
+  public handleClickLogin = (e: any) => {
+    enableMetamask();
+  }
+
   public render() {
-    let {
-      currentAccount,
+    const {
       currentAccountProfile,
       dao,
       ethAccountAddress,
-      networkId,
-      pageURL,
     } = this.props;
 
     const daoAvatarAddress = dao ? dao.address : null;
-
-    const isProfilePage = pageURL.includes("profile");
 
     return(
       <div>
@@ -101,14 +94,18 @@ class HeaderContainer extends React.Component<IProps, null> {
             { ethAccountAddress &&
               <div className={css.accountInfo}>
                 <div className={css.accountImage}>
-                  <Link className={css.profileLink} to={"/profile/" + ethAccountAddress + (daoAvatarAddress ? "?daoAvatarAddress=" + daoAvatarAddress : "")}>
+                  <Link className={css.profileLink}
+                    to={"/profile/" + ethAccountAddress + (daoAvatarAddress ? "?daoAvatarAddress=" + daoAvatarAddress : "")}>
                     <AccountImage accountAddress={ethAccountAddress} />
                   </Link>
                 </div>
                 <div className={css.holdings}>
                   <div className={css.pointer}></div>
                   <div className={css.walletDetails}>
-                    <div className={css.profileName}><AccountProfileName accountAddress={ethAccountAddress} accountProfile={currentAccountProfile} daoAvatarAddress={daoAvatarAddress} /></div>
+                    <div className={css.profileName}>
+                      <AccountProfileName accountAddress={ethAccountAddress}
+                        accountProfile={currentAccountProfile} daoAvatarAddress={daoAvatarAddress} />
+                    </div>
                     <div className={css.holdingsLabel}>Your wallet</div>
                     <div className={css.copyAddress} style={{cursor: "pointer"}} onClick={this.copyAddress}>
                       <span>{ethAccountAddress ? ethAccountAddress.slice(0, 40) : "No account known"}</span>
@@ -121,13 +118,11 @@ class HeaderContainer extends React.Component<IProps, null> {
               </div>
             ||
               <div className={css.accountInfo}>
-                Please log in!
+                <button onClick={this.handleClickLogin}>
+                  Please log in!
+                </button>
               </div>
             }
-            {/* dao && !isProfilePage
-              ? <button className={css.openTour} onClick={this.handleClickTour}><img src="/assets/images/Tour/TourButton.svg"/></button>
-              : ""
-            */}
           </div>
         </nav>
       </div>
